+++ src/ann/__init__.py
@@ -1,3 +1,0 @@
-"""Ann core package."""
-
-__version__ = "0.0.2"